+++ frontend/src/services/adminIngredientApi.ts
@@ -1,32 +1,36 @@
-import apiClient from "./adminApi" //
-import type { Ingredient, IngredientView } from "../model/types"
+import apiClient from './adminApi'; //
+import type {Ingredient, IngredientView} from '../model/types';
 
 export const addIngredient = async (ingredient: Ingredient): Promise<void> => {
   try {
-    await apiClient.post("/admin/ingredient/addIngredient", ingredient)
+    await apiClient.post('/admin/ingredient/addIngredient', ingredient);
   } catch (error) {
-    throw new Error("Ошибка добавления ингредиента")
+    throw new Error('Ошибка добавления ингредиента');
   }
-}
+};
 
 export const getAllIngredients = async (): Promise<IngredientView[]> => {
   try {
-    const response = await apiClient.get("/admin/ingredient/getAllIngredient")
-    return response.data
+    const response = await apiClient.get('/admin/ingredient/getAllIngredient');
+    return response.data;
   } catch (error) {
-    throw new Error("Ошибка при получении данных")
+    throw new Error('Ошибка при получении данных');
   }
-}
+};
 
-export const deleteIngredient = async (id: number): Promise<void> => {
+export const updateIngredient = async (id: number, data: { name?: string }) => {
+  const response = await apiClient.put(`/admin/ingredient/updateIngredient/${id}`, data);
+  return response.data;
+};
+
+export const deleteIngredient = async (id: number): Promise<boolean> => {
   try {
-    const response = await apiClient.delete(
-      `/admin/ingredient/deleteIngredient/${id}`
-    )
-    if (response.status !== 200) {
-      throw new Error("Ошибка удаления")
+    const response = await apiClient.delete(`/admin/ingredient/deleteIngredient/${id}`);
+    return response.status === 200;
+  } catch (err: any) {
+    if (err.response?.status === 409) {
+      throw err.response.data;
     }
-  } catch (error) {
-    throw new Error("Ошибка удаления")
+    return false;
   }
-}
+};
